requirements.txt
python-socketio
websocket-client
python-dotenv
loguru
